mavu-core/mavu_fullstack-main/backend/requirements.txt
# Core dependencies
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
websockets==13.1

# OpenAI and ML
openai==2.8.1
tiktoken==0.8.0

# PostgreSQL Database
sqlalchemy==2.0.35
asyncpg==0.29.0
psycopg2-binary==2.9.10
alembic==1.13.3

# Weaviate Vector DB
weaviate-client==4.9.4

# Redis for caching and chat history
redis[asyncio]==5.2.0
hiredis==3.0.0

# Data processing
numpy==1.26.4
pydantic==2.9.2  # Keep <2.10 for aiogram compatibility
pydantic-settings==2.6.0

# Utils
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.1
aiofiles==24.1.0
tenacity==8.2.3

# CLI tool
click==8.1.7

# Rate limiting
slowapi==0.1.9

# Logging
structlog==24.4.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.0

# Admin Interface
sqladmin[full]>=0.16.0
itsdangerous==2.1.2
markupsafe==2.1.5

# Telegram Bot
aiogram==3.13.1
aiohttp==3.10.10